% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_install.R
\name{check_install_package}
\alias{check_install_package}
\title{Check for and optionally install packages}
\usage{
check_install_package(x, auto.install = FALSE)
}
\arguments{
\item{x}{The package to be checked}

\item{auto.install}{A logical indicating whether to 
automatically install missing packages or check with user first}
}
\value{
Message indicating status of installation
}
\description{
checks to make sure package is installed. If not, 
either auto installs or checks with user first
}
