% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_padrino.R
\name{predict_Padrino}
\alias{predict_Padrino}
\title{Generic PADRINO parameter extraction function}
\usage{
predict_Padrino(Package, Class, Model, MeshPts, lower = NULL, upper = NULL)
}
\arguments{
\item{Package}{The package housing the predict method to be used}

\item{Class}{The class of the model to be predicted}

\item{Model}{A model object that contains, at a minimum, the names
and estimates of the response variable, fixed and random effects.}

\item{MeshPts}{The meshpoints of the IPM matrix to be created.}

\item{lower}{In the case of a model with a single fixed effect,
a numeric indicating the lower bound of the range of the state variable
in the regression. In the case of models with multiple fixed effects,
a named list with where each component is a numeric indicating the lower
bound of the effect.}

\item{upper}{In the case of a model with a single fixed effect,
a numeric indicating the upper bound of the range of the state variable
in the regression. In the case of models with multiple fixed effects,
a named list with where each component is a numeric indicating the upper
bound of the effect.}
}
\description{
Generic PADRINO parameter extraction function
}
